refactor: 프로젝트 구조 및 충돌 해결
