+++ build.gradle
@@ -21,6 +21,9 @@ repositories {
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation 'org.hibernate.validator:hibernate-validator:6.2.0.Final'
+
     compileOnly 'org.projectlombok:lombok'
 //	DB 설정 후 build
     runtimeOnly 'com.mysql:mysql-connector-j'
