+++ src/main/java/com/hangout/hangout/HangoutApplication.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @EnableJpaAuditing
 @SpringBootApplication
-@EnableJpaAuditing
 public class HangoutApplication {
 
 	public static void main(String[] args) {
