+++ src/main/java/com/hangout/hangout/domain/report/entity/PostReport.java
@@ -2,8 +2,8 @@ package com.hangout.hangout.domain.report.entity;
 
 import com.hangout.hangout.domain.post.entity.Post;
 import com.hangout.hangout.domain.user.entity.User;
-import com.hangout.hangout.global.common.domain.BaseEntity;
-import com.hangout.hangout.global.common.domain.Status;
+import com.hangout.hangout.global.common.domain.entity.BaseEntity;
+import com.hangout.hangout.global.common.domain.entity.Status;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
